drm/vc4: Make sure that waits that get interrupted don't wait forever.
danvet's talk recommended avoiding relative waits entirely, but his
example was of relative waits with units (vblank numbers) that were
potentially longer than the interval between signals coming in, so you
couldn't adjust the ioctl arguments appropriately.

This change also has a wait unit that might be too large (jiffies),
but given that the other argument (an emitted seqno) is guaranteed to
eventually result in a nonblocking successful return, this isn't
really a big deal.

Signed-off-by: Eric Anholt <eric@anholt.net>