ARM: bcm2835: Add the mailbox property channel driver.
Many of the operations with the firmware are done through this mailbox
channel pair with a specific packet format.  Notably, it's used for
clock control, which is apparently not actually totally possible to do
from the ARM side (some regs aren't addressable).  I need clock
control for the VC4 DRM driver, to turn on the 3D engine.

v2: Move to arch/arm/mach-bcm, since it's specific to the
    architecture's firmware, not the mailbox device.

v3: Fix send_message arg for the change to pass a pointer to our data.
    Fix returning -EPROBE_DEFER when called before the module is
    probed, and fix deferring our own probe when the channel isn't
    ready.  Drop extra error message.  Don't lose return value from
    bcm_mbox_property_transaction().

Signed-off-by: Eric Anholt <eric@anholt.net>